app/requirements.txt
uvicorn
fastapi
SQLAlchemy
alembic
python-dotenv
pydantic
psycopg2
celery